Fixed test_get_single_object_by_id: removed unnecessary checks. Fixed the test_delete_object test: moved assertions to the methods of the corresponding classes. Fixed the structure of the project: moved each class into a separate file.
