+++ test_api_svetlanagavs/controllers/base_data_object.py
@@ -1,0 +1,7 @@
+class ObjectData:
+    def __init__(self, name, year, price, cpu_model, hard_disk_size):
+        self.name = name
+        self.year = year
+        self.price = price
+        self.cpu_model = cpu_model
+        self.hard_disk_size = hard_disk_size
