+++ test_api_svetlanagavs/tests/test_object.py
@@ -13,21 +13,20 @@ def test_get_list_by_ids(get_list_by_ids):
     get_list_by_ids.check_object_is_not_empty()
 
 
-@pytest.mark.parametrize('add_object', [('Apple MacBook Pro 14 get id', 2022, 2000, 'i7', '1TB', True)], indirect=True)
-def test_get_single_object_by_id(add_object, get_single_object_by_id):
-    created_object, test_data = add_object
+@pytest.mark.parametrize(
+    'add_delete_object', [('Apple MacBook Pro 14 get id', 2022, 2000, 'i7', '1TB', True)], indirect=True
+)
+def test_get_single_object_by_id(add_delete_object, get_single_object_by_id):
+    created_object, test_data = add_delete_object
     object_id = created_object.json['id']
 
     get_single_object_by_id.get_single_object_by_id(object_id)
     get_single_object_by_id.check_status_code()
     get_single_object_by_id.check_id(object_id)
-    get_single_object_by_id.check_instance(dict)
-    get_single_object_by_id.check_object_is_not_empty()
-    get_single_object_by_id.check_object_name(test_data.name)
 
 
 @pytest.mark.parametrize(
-    'add_object',
+    'add_delete_object',
     [
         ('Apple MacBook Pro 14 test', 2022, 2000, 'i7', '1TB', True),
         ('Apple MacBook Pro 15 test', 2023, 2500, 'i8', '2TB', True),
@@ -35,8 +34,8 @@ def test_get_single_object_by_id(add_object, get_single_object_by_id):
     ],
     indirect=True
 )
-def test_add_object(add_object):
-    created_object, test_data = add_object
+def test_add_object(add_delete_object):
+    created_object, test_data = add_delete_object
     object_id = created_object.json['id']
 
     created_object.check_status_code()
@@ -49,9 +48,11 @@ def test_add_object(add_object):
     )
 
 
-@pytest.mark.parametrize('add_object', [('Apple MacBook Pro 14 update', 2022, 2000, 'i7', '1TB', False)], indirect=True)
-def test_update_object(add_object, update_object):
-    created_object = add_object
+@pytest.mark.parametrize(
+    'add_delete_object', [('Apple MacBook Pro 14 update', 2022, 2000, 'i7', '1TB', False)], indirect=True
+)
+def test_update_object(add_delete_object, update_by_id_object):
+    created_object = add_delete_object
     payload_update = created_object.json
 
     update_id = payload_update['id']
@@ -60,20 +61,20 @@ def test_update_object(add_object, update_object):
     payload_update['data']['price'] = 3600
     payload_update['data']['color'] = 'silver'
 
-    update_object.update_object(update_id, payload_update)
-    update_object.check_status_code()
-    update_object.check_id(update_id)
-    update_object.check_instance(dict)
-    update_object.check_object_is_not_empty()
-    update_object.check_object_name(payload_update['name'])
-    update_object.check_object_entire_data(payload_update['data'])
+    update_by_id_object.update_object(update_id, payload_update)
+    update_by_id_object.check_status_code()
+    update_by_id_object.check_id(update_id)
+    update_by_id_object.check_instance(dict)
+    update_by_id_object.check_object_is_not_empty()
+    update_by_id_object.check_object_name(payload_update['name'])
+    update_by_id_object.check_object_entire_data(payload_update['data'])
 
 
 @pytest.mark.parametrize(
-    'add_object', [('Apple MacBook Pro 14 partial update', 2022, 2000, 'i7', '1TB', False)], indirect=True
+    'add_delete_object', [('Apple MacBook Pro 14 partial update', 2022, 2000, 'i7', '1TB', False)], indirect=True
 )
-def test_update_partial_object(add_object, partial_update_object):
-    created_object = add_object
+def test_update_partial_object(add_delete_object, partial_update_object):
+    created_object = add_delete_object
     payload_update = created_object.json
     # update partial object
     payload_partial = {"name": "Apple MacBook Pro 16 (Updated Name PARTIAL)"}
@@ -88,18 +89,16 @@ def test_update_partial_object(add_object, partial_update_object):
     partial_update_object.check_object_entire_data(payload_update['data'])
 
 
-@pytest.mark.parametrize('add_object', [('Apple MacBook Pro 14 delete', 2022, 2000, 'i7', '1TB', False)], indirect=True)
-def test_delete_object(add_object, get_single_object_by_id, delete_object):
-    created_object = add_object
+@pytest.mark.parametrize(
+    'add_delete_object', [('Apple MacBook Pro 14 delete', 2022, 2000, 'i7', '1TB', False)], indirect=True
+)
+def test_delete_object(add_delete_object, get_single_object_by_id, delete_by_id_object):
+    created_object = add_delete_object
     object_id = created_object.json['id']
 
-    delete_object.delete_object(object_id)
-    delete_object.check_status_code()
-
-    assert delete_object.json['message'] == f'Object with id = {object_id} has been deleted.', \
-        f"Message is incorrect: {delete_object.json['message']}\n{delete_object.curl}"
-
+    delete_by_id_object.delete_object(object_id)
+    delete_by_id_object.check_status_code()
+    delete_by_id_object.check_delete_message(object_id)
     get_single_object_by_id.get_single_object_by_id(object_id)
     get_single_object_by_id.check_status_code(404)
-    assert get_single_object_by_id.json['error'] == f'Oject with id={object_id} was not found.', \
-        f'Object was not deleted\n{get_single_object_by_id.curl}'
+    get_single_object_by_id.check_error_message(object_id)
